SFIR-457 Fix the way we log info in our application (#162)
Pino expects the log methods to be called with `(obj, string)`
not `(string, obj)`.

Also the CDP Support OpenSearch Log ingester is not consuming our
bespoke JSON logs as they don't conform with the default keys they
are consuming from the Pino method.

Therefore we have to do string interpolation and stringify our JSON
payload in order to have it visible in CDP OpenSearch Logs